+++ shellcheck.sh
@@ -44,31 +44,40 @@ current_version=$(find /opt -maxdepth 1 -type d -name "shellcheck *" -printf '%f
 #Start installation if github version is not equal to installed version
 if [ "$tag_name" != "$current_version" ] && [ $checkFlag = false ] || [ $forceFlag = true ]
 then
+  printf "Begin %s installation..." "$program"
+
   #Download binaries
   curl -s https://api.github.com/repos/koalaman/shellcheck/releases/latest \
   | grep "browser_download_url.*linux.x86_64.tar.xz\"" \
   | cut -d \" -f 4 \
-  | xargs curl -L -o /tmp/shellcheck.tar.xz
+  | xargs curl -Lsf -o /tmp/shellcheck.tar.xz
 
-  #Remove contents if already installed
-  find /opt -maxdepth 1 -type d -name "shellcheck *" -exec sudo rm -rf '{}' \+
+  if [ $? = 0 ]
+  then
+    #Remove contents if already installed
+    find /opt -maxdepth 1 -type d -name "shellcheck *" -exec sudo rm -rf '{}' \+
 
-  #Create folder for contents
-  sudo mkdir -p "$installDir"
+    #Create folder for contents
+    sudo mkdir -p "$installDir"
 
-  #Expand tar file to folder
-  sudo tar Jxf /tmp/shellcheck.tar.xz --strip-components=1 -C "$installDir"
+    #Expand tar file to folder
+    sudo tar Jxf /tmp/shellcheck.tar.xz --strip-components=1 -C "$installDir"
 
-  #Change execute permissions
-  sudo chmod +x "$installDir/shellcheck"
+    #Change execute permissions
+    sudo chmod +x "$installDir/shellcheck"
 
-  #Create symbolic link to bin folder
-  sudo mkdir -p /usr/local/bin
-  sudo ln -sf "$installDir/shellcheck" /usr/local/bin
+    #Create symbolic link to bin folder
+    sudo mkdir -p /usr/local/bin
+    sudo ln -sf "$installDir/shellcheck" /usr/local/bin
+
+    printf "Finished\n"
+  else
+    printf "Failed\n"
+  fi
 
 elif [ $checkFlag = true ] && [ "$tag_name" = "$current_version" ]
 then
-  echo "Update not found for $program"
+  echo "No update found for $program"
 
 elif [ $checkFlag = true ] && [ "$tag_name" != "$current_version" ]
 then
